controller: Merge the router database into the controller database
