+++ bootstrap/manifest_template.json
@@ -469,7 +469,12 @@
         "TLSCERT": "{{ (index .StepData \"controller-cert\").Cert }}",
         "TLSKEY": "{{ (index .StepData \"controller-cert\").PrivateKey }}",
         "COOKIE_KEY": "{{ (index .StepData \"router-sticky-key\").Data }}",
-        "PROXY_PROTOCOL": "{{ getenv \"PROXY_PROTOCOL\" }}"
+        "PROXY_PROTOCOL": "{{ getenv \"PROXY_PROTOCOL\" }}",
+        "FLYNN_POSTGRES": "{{ (pgconfig .StepData \"controller\").ServiceName }}",
+        "PGHOST": "{{ (pgconfig .StepData \"controller\").Host }}",
+        "PGUSER": "{{ (pgconfig .StepData \"controller\").User }}",
+        "PGPASSWORD": "{{ (pgconfig .StepData \"controller\").Password }}",
+        "PGDATABASE": "{{ (pgconfig .StepData \"controller\").Database }}"
       },
       "processes": {
         "app": {
@@ -482,8 +487,7 @@
     },
     "processes": {
       "app": 1
-    },
-    "resources": [{"name":"postgres", "url":"http://postgres-api.discoverd/databases"}]
+    }
   },
   {
     "id": "gitreceive",
