+++ router/server.go
@@ -206,16 +206,7 @@ func main() {
 	}
 
 	log.Info("connecting to postgres")
-	db := postgres.Wait(nil, nil)
-
-	log.Info("running DB migrations")
-	if err := migrateDB(db); err != nil {
-		shutdown.Fatal(err)
-	}
-	db.Close()
-
-	log.Info("reconnecting to postgres with prepared queries")
-	db = postgres.Wait(nil, schema.PrepareStatements)
+	db := postgres.Wait(nil, schema.PrepareStatements)
 
 	shutdown.BeforeExit(func() { db.Close() })
